Revert "Added some models"
